requirements.txt
qiskit
cirq-core
openfermion
qiskit-addon-sqd
